Fix create method compliance #1518
