+++ Physical_Engine/Create/Column.cs
@@ -1,4 +1,4 @@
-/*
+﻿/*
  * This file is part of the Buildings and Habitats object Model (BHoM)
  * Copyright (c) 2015 - 2020, the respective contributors. All rights reserved.
  *
@@ -35,18 +35,6 @@ namespace BH.Engine.Physical
         /**** Public Methods                            ****/
         /***************************************************/
 
-        [Description("Creates a physical Beam element. To generate elements compatible with structural packages, have a look at the Bar class")]
-        [Input("location", "The centre line geometry of the beam")]
-        [Input("property", "The property of the beam, containing its profile, orientation and materiality")]
-        [Input("name", "The name of the beam, default empty string")]
-        [Output("Beam", "The created physical beam")]
-        public static Beam Beam(ICurve location, IFramingElementProperty property, string name = "")
-        {
-            return new Beam { Location = location, Property = property, Name = name };
-        }
-
-        /***************************************************/
-
         [Description("Creates a physical Column element. To generate elements compatible with structural packages, have a look at the Bar class")]
         [Input("location", "The centre line geometry of the Column")]
         [Input("property", "The property of the column, containing its profile, orientation and materiality")]
@@ -68,44 +56,7 @@ namespace BH.Engine.Physical
         {
             return Create.Column(Geometry.Create.Line(basePoint, basePoint + Vector.ZAxis * height), property, name);
         }
-
-        /***************************************************/
-
-        [Description("Creates a physical Bracing element. To generate elements compatible with structural packages, have a look at the Bar class")]
-        [Input("location", "The centre line geometry of the Bracing")]
-        [Input("property", "The property of the bracing, containing its profile, orientation and materiality")]
-        [Input("name", "The name of the bracing, default empty string")]
-        [Output("Bracing", "The created physical Bracing")]
-        public static Bracing Bracing(ICurve location, IFramingElementProperty property, string name = "")
-        {
-            return new Bracing { Location = location, Property = property, Name = name };
-        }
-
-        /***************************************************/
-
-        [Description("Creates a physical Cable element. To generate elements compatible with structural packages, have a look at the Bar class")]
-        [Input("location", "The centre line geometry of the Cable")]
-        [Input("property", "The property of the cable, containing its profile, orientation and materiality")]
-        [Input("name", "The name of the cable, default empty string")]
-        [Output("Cable", "The created physical Cable")]
-        public static Cable Cable(ICurve location, IFramingElementProperty property, string name = "")
-        {
-            return new Cable { Location = location, Property = property, Name = name };
-        }
-
-        /***************************************************/
-
-        [Description("Creates a physical Pile element. To generate elements compatible with structural packages, have a look at the Bar class")]
-        [Input("location", "The centre line geometry of the Pile")]
-        [Input("property", "The property of the pile, containing its profile, orientation and materiality")]
-        [Input("name", "The name of the pile, default empty string")]
-        [Output("Pile", "The created physical Pile")]
-        public static Pile Pile(ICurve location, IFramingElementProperty property, string name = "")
-        {
-            return new Pile { Location = location, Property = property, Name = name };
-        }
-
-        /***************************************************/
+ 
     }
 }
