+++ Physical_Engine/Physical_Engine.csproj
@@ -64,10 +64,14 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Create\Bracing.cs" />
+    <Compile Include="Create\Cable.cs" />
+    <Compile Include="Create\Column.cs" />
+    <Compile Include="Create\Pile.cs" />
     <Compile Include="Create\Roof.cs" />
-    <Compile Include="Create\FramingElement.cs" />
+    <Compile Include="Create\Beam.cs" />
     <Compile Include="Create\Construction.cs" />
-    <Compile Include="Create\FramingElementProperty.cs" />
+    <Compile Include="Create\ConstantFramingProperty.cs" />
     <Compile Include="Create\Floor.cs" />
     <Compile Include="Create\Wall.cs" />
     <Compile Include="Create\Layer.cs" />
